+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/configuration/arena/modules/MissileConfig.java
@@ -16,7 +16,7 @@
  * along with MissileWars.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.butzlabben.missilewars.configuration.arena;
+package de.butzlabben.missilewars.configuration.arena.modules;
 
 import de.butzlabben.missilewars.Logger;
 import de.butzlabben.missilewars.game.schematics.SchematicConfiguration;
