+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/configuration/arena/modules/MoneyConfig.java
@@ -16,9 +16,8 @@
  * along with MissileWars.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.butzlabben.missilewars.configuration.arena;
+package de.butzlabben.missilewars.configuration.arena.modules;
 
-import com.google.gson.annotations.SerializedName;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.ToString;
@@ -26,8 +25,9 @@ import lombok.ToString;
 @Getter
 @ToString
 @RequiredArgsConstructor
-public class GameRespawnConfig {
+public class MoneyConfig {
 
-    @SerializedName("send_bow") private boolean sendBow = true;
-    @SerializedName("send_pickaxe") private boolean sendPickaxe = false;
+    private int win = 80;
+    private int loss = 50;
+    private int draw = 30;
 }
